requirements.txt
flask
python-dotenv
requests
google-generativeai
edge-tts
ffmpeg-python
moviepy
groq
duckduckgo-search
Pillow
